4.21 tokenized deleting

+++ osa4/blogilista/controllers/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const { usersInDB } = require('../tests/test_helper')
 
 blogsRouter.get('/', async (request, response) => {
 	//const blogs = await Blog.find({})
@@ -44,9 +45,29 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+	
+	const blog = await Blog.findById(request.params.id)
+	console.log('id', request.params.id)
+	console.log('blog', blog)
+	
+	
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid'} )
+	}
+	const user = await User.findById(decodedToken.id)
+	console.log('user', user)
+	console.log('user_id', user._id.toString())
+	console.log('blog.user.id', blog.user.toString())
+	
+	
+	if (user._id.toString() === blog.user.toString()) {
+		await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).end()
+	}
+	else {
+		response.status(401).json( {error: 'no blog found' })
+	}
 })
 
 blogsRouter.put('/:id', async (request, response) => {
